fix: increasing CPU load due to libmav queue
libmav has an unordered map for the expected messages. When the receive was called with a timeout
and the timeout was reached, then the expected message was not removed from the map.
The CAMERA_SETTINGS message is only send when the settings change or on request, so most of the
calls to receive timed out. As a result, the number of items in the map kept increasing which in
turn lead to the increase in CPU load.
This commit updates the libmav submodule to a version with the fix.