fix: FPV camera used for scaling in multi-payload setups
Astro is enabling support for an FPV camera in addition to the main payload. The current
implementation in the slowmode-app uses the first camera that shows up which most often is the
FPV camera. In that case, the yaw rate scaling based on the zoom doesn't work.
On Astro, A7R and Wiris are assigned the component ID 100. So, for now we limit the range of valid
component IDs to just 100.